[jit] make containedElements a vector
The bitset doesn't preserve ordering, which will be important for doing
contained element analysis on tuples/classes.

One annoying thing: If we have a tuple like `(Tensor, int, Tensor)`, we
want the contained elements to match, so that indexing into
`type->containedTypes()` is the same as indexing into
`el->containedElements()`. But since `int` is not mutable, we need to
make the vector `(Element*, nullptr, Element*)`. Open to suggestions how
to avoid.

gh-metadata: pytorch pytorch 21445 gh/suo/54/head