Remove the misguided attempt at TORCH_STATIC.
TORCH_STATIC does something, but it's not what you want.  What the
TORCH_STATIC toggle does is cause libtorch to be generated as a static
library, rather than a dynamic library.  This sounds useful, until you
realize that it doesn't affect any of the *other* libraries we also
define in PyTorch, e.g., libcaffe2 or libc10, etc. etc.  So you end up
with one library that is static... that links to a pile of dynamic
libraries.  Not useful.

The proper way to build everything "static" is to set -DBUILD_SHARED_LIBS=OFF
which will turn the interpretation of any add_library() specifier to
generate a static library (assuming you didn't override it with SHARED
which is what happened here).

There is some attempt at building static libraries in builder, but it
doesn't seem to work, and in any case we don't seem to publish them.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

gh-metadata: pytorch pytorch 21704 gh/ezyang/168/head