Avoid weak_ptr::lock() when shared_ptr is provably live; more error checking.
In the original iteration of the patch, I used lock() everywhere to minimize
the amount of code I have to modify.  In this patch, I now eliminate as many
lock()s as I can, when the caller is known to have a strong reference to the
PyFunction, and pass that directly.

Along the way, I also bulk up our error messages for checking the result
of the weak pointer dereference.  Some of these cases can be triggered
by zany use of legacy autograd function API; might as well let people know
what they've done wrong.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 296f2e6edea9e426201828ec1adb5f28fb351d5f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/22998