[jit] make the optimization flag more consistent
1. Optimizations were previosuly set per-compilation unit. Now that
there can be multiple modules in a single CU, I changed it to be a table
of qualified name => flag. In the long run this should just be a regular
attribute of the class type, but I just want to unblock things for now.
2. Pass the optimization flag through @script more consistenly

gh-metadata: pytorch pytorch 23153 gh/suo/101/head